agent/requirements.txt
strands-agents
strands-agents-builder
boto3
